+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## [1.0.24] - 2021-01-02
+
+### Added
+- add option `invert` for `PLC_LockMechanismBool`
+- add option `mapGet` and `mapSet` for `PLC_LockMechanism`
+
 
 ## [1.0.23] - 2021-01-02